+++ packages/action/src/api.ts
@@ -1,5 +1,5 @@
 import { createHash } from 'crypto';
-import { BaseApiClient } from '@openzeppelin/platform-sdk-base-client';
+import { BaseApiClient } from '@openzeppelin/defender-sdk-base-client';
 import { CreateActionRequest, UpdateActionRequest, GetSecretsResponse, SaveSecretsRequest } from './models/action';
 import { ActionRunBase, ActionRunListResponse, ActionRunResponse, ActionRunStatus } from './models/action-run.res';
 import { ActionDeleteResponse, ActionListResponse, ActionResponse } from './models/response';
@@ -12,16 +12,16 @@ type SourceFiles = {
 
 export class ActionClient extends BaseApiClient {
   protected getPoolId(): string {
-    return process.env.PLATFORM_POOL_ID || 'us-west-2_94f3puJWv';
+    return process.env.DEFENDER_POOL_ID || 'us-west-2_94f3puJWv';
   }
 
   protected getPoolClientId(): string {
-    return process.env.PLATFORM_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
+    return process.env.DEFENDER_POOL_CLIENT_ID || '40e58hbc7pktmnp9i26hh5nsav';
   }
 
   protected getApiUrl(): string {
-    // TODO: update to platform-api.* url (and /action) when available
-    return process.env.PLATFORM_API_URL || 'https://defender-api.openzeppelin.com/autotask/';
+    // TODO: update to /action when available
+    return process.env.DEFENDER_API_URL || 'https://platform-api.openzeppelin.com/autotask/';
   }
 
   public async list(): Promise<ActionListResponse> {