+++ packages/deploy/src/models/deployment-config.ts
@@ -1,4 +1,4 @@
-import { Network } from '@openzeppelin/platform-sdk-base-client';
+import { Network } from '@openzeppelin/defender-sdk-base-client';
 
 export interface DeploymentConfigCreateRequest {
   relayerId: string;