+++ packages/proposal/src/models/proposal.ts
@@ -1,4 +1,4 @@
-import { Network } from '@openzeppelin/platform-sdk-base-client';
+import { Network } from '@openzeppelin/defender-sdk-base-client';
 
 // Copied from openzeppelin/defender/models/src/types/proposal-api.req.d.ts
 export type Hex = string;