+++ packages/notification-channel/CHANGELOG.md
@@ -1,15 +1,26 @@
-# @openzeppelin/platform-sdk-notification-channel-client
+# @openzeppelin/defender-sdk-notification-channel-client
+
+## 0.4.0
+
+### Minor Changes
+
+- be73034: Feat: Support linea & base
+
+### Patch Changes
+
+- Updated dependencies [be73034]
+  - @openzeppelin/platform-sdk-base-client@0.4.0
 
 ## 0.3.0
 
 ### Minor Changes
 
-- efefd68: feat: Add support of instantiating platform-sdk package from Platform Actions
+- efefd68: feat: Add support of instantiating defender-sdk package from Platform Actions
 
 ### Patch Changes
 
 - Updated dependencies [efefd68]
-  - @openzeppelin/platform-sdk-base-client@0.3.0
+  - @openzeppelin/defender-sdk-base-client@0.3.0
 
 ## 0.2.0
 
@@ -28,4 +39,4 @@
 ### Patch Changes
 
 - Updated dependencies [496f88d]
-  - @openzeppelin/platform-sdk-base-client@0.2.0
+  - @openzeppelin/defender-sdk-base-client@0.2.0