+++ packages/relay-signer/CHANGELOG.md
@@ -1,15 +1,26 @@
-# @openzeppelin/platform-sdk-relay-signer-client
+# @openzeppelin/defender-sdk-relay-signer-client
+
+## 0.4.0
+
+### Minor Changes
+
+- be73034: Feat: Support linea & base
+
+### Patch Changes
+
+- Updated dependencies [be73034]
+  - @openzeppelin/platform-sdk-base-client@0.4.0
 
 ## 0.3.0
 
 ### Minor Changes
 
-- efefd68: feat: Add support of instantiating platform-sdk package from Platform Actions
+- efefd68: feat: Add support of instantiating platform-sdk package from Defender Actions
 
 ### Patch Changes
 
 - Updated dependencies [efefd68]
-  - @openzeppelin/platform-sdk-base-client@0.3.0
+  - @openzeppelin/defender-sdk-base-client@0.3.0
 
 ## 0.2.0